requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.0
httpx>=0.26.0
pytest==7.4.3
gunicorn>=23.0.0
langgraph>=0.2.66
langchain>=0.3.0
langchain-openai>=0.2.0
supabase>=2.27.0
postgrest>=2.27.0
gotrue>=2.12.4
google-api-python-client>=2.111.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
newrelic>=9.0.0
